refactor(EG-941): run details and upload step components into overall stepper component
